// packages/frontend/e2e-tests/cypress/lib/interfaces/trading-servers-manager/taps-interfaces.ts
export enum ETabUrl {
    addTask = '?tab=Add%20Task',
    instruments = '?tab=Instruments',
    indicators = '?tab=Indicators',
    productLogs = '?tab=Product%20Logs',
    realAccounts = '?tab=Real%20Accounts',
    virtualAccounts = '?tab=Virtual%20Accounts',
    wsRequest = '?tab=WS%20Request',
    orderBook = '?tab=Order%20Book',
    config = '?tab=Config',
    dashboards = '?tab=Dashboards',
    statusMessages = '?tab=Status%20Messages',
    activeOrders = '?tab=Active%20Orders',
    addComponent = '?tab=Add%20Component',
    state = '?tab=State',
    balances = '?tab=Balances',
    positions = '?tab=Positions',
    robotBalances = '?tab=Robot%20Balances',
    robotPositions = '?tab=Robot%20Positions',
}